feat: CreateTables utility
